refactor: 패키지 네이밍 및 구조 변경
